+++ src/main/java/com/epde/rt/model/tasks/enums/TaskPriority.java
@@ -1,0 +1,9 @@
+package com.epde.rt.model.tasks.enums;
+
+public enum TaskPriority {
+    Low,
+    Normal,
+    High,
+    Urgent,
+    Immediate
+}
