+++ src/main/java/com/epde/rt/model/tasks/enums/TaskStatus.java
@@ -1,0 +1,7 @@
+package com.epde.rt.model.tasks.enums;
+
+public enum TaskStatus {
+    New,
+    Processing,
+    Completed
+}
