+++ src/main/java/com/epde/rt/repository/TasksRepository.java
@@ -1,6 +1,6 @@
 package com.epde.rt.repository;
 
-import com.epde.rt.model.Tasks;
+import com.epde.rt.model.tasks.Tasks;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
