+++ src/main/resources/templates/tasks/task_index.html
@@ -14,7 +14,7 @@
 <div class="container my-2">
     <h1 align="center">Tasks List</h1>
 
-    <a th:href="@{/showNewTasksForm}" class="btn btn-primary btn-sm mb-3"> Add Task </a>
+    <a th:href="@{/showNewTasksForm}" class="btn btn-primary">Add Task</a>
 
     <table border="1" class = "table table-striped table-responsive-md" align="center">
         <thead align="center">
@@ -42,6 +42,7 @@
             </th>
         </tr>
         </thead>
+
         <tbody align="center">
         <tr th:each="tasks : ${listTasks}">
             <td>
@@ -79,5 +80,6 @@
         </div>
     </div>
 </div>
+
 </body>
 </html>
