+++ src/main/resources/templates/tasks/update_tasks.html
@@ -5,7 +5,9 @@
     <meta charset="ISO-8859-1">
     <title>Task Application System</title>
 
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css" integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO" crossorigin="anonymous">
+    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/css/bootstrap.min.css"
+          integrity="sha384-MCw98/SFnGE8fJT3GXwEOngsV7Zt27NXFoaoApmYm81iuXoPkFOJwJ8ERdknLPMO"
+          crossorigin="anonymous">
 </head>
 
 <body>
@@ -16,31 +18,52 @@
 
     <form action="#" th:action="@{/saveTasks}" th:object="${tasks}" method="POST">
 
-        <!-- Add hidden form field to handle update -->
         <input type="hidden" th:field="*{taskId}" />
+        <div class="form-group">
+            <label></label>
+            <input type="hidden" th:field="*{taskId}" />
+        </div>
 
-        <label>
-            <input type="text" th:field="*{taskTitle}" class="form-control mb-4 col-4">
-        </label>
+        <div class="form-group">
+            <label>Title</label>
+            <input type="text" th:field="*{taskTitle}" placeholder="Title" class="form-control">
+        </div>
 
-        <label>
-            <input type="text" th:field="*{taskDetails}" class="form-control mb-4 col-4">
-        </label>
+        <div class="form-group">
+            <label>Details</label>
+            <input type="text" th:field="*{taskDetails}" placeholder="Details" class="form-control">
+        </div>
 
-        <label>
-            <input type="text" th:field="*{taskPriority}" class="form-control mb-4 col-4">
-        </label>
+        <div class="form-group">
+            <label>Priority</label>
+            <select class="form-control" th:field="*{taskPriority}" id="taskPriority">
+                <option value="">Select Priority</option>
+                <option th:each="taskPriority : ${T(com.epde.rt.model.tasks.enums.TaskPriority).values()}"
+                        th:value="${taskPriority}"
+                        th:text="${taskPriority}">
+                    Task Priority
+                </option>
+            </select>
+        </div>
 
-        <label>
-            <input type="text" th:field="*{taskStatus}" class="form-control mb-4 col-4">
-        </label>
+        <div class="form-group">
+            <label>Status</label>
+            <select class="form-control" th:field="*{taskStatus}" id="taskStatus">
+                <option value="">Select Status</option>
+                <option th:each="taskStatus : ${T(com.epde.rt.model.tasks.enums.TaskStatus).values()}"
+                        th:value="${taskStatus}"
+                        th:text="${taskStatus}">
+                    Task Status
+                </option>
+            </select>
+        </div>
 
-        <button type="submit" class="btn btn-info col-2"> Update Task</button>
+        <button type="submit" class="btn btn-primary">Update</button>
     </form>
 
     <hr>
 
-    <a th:href="@{/static}"> Back to Tasks List</a>
+    <a th:href="@{/}"> Back to Tasks List</a>
 </div>
 </body>
